Removes RoadGeometryBuilderBase in favor of less (unnecessary) abstraction.
